requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.2
Django==3.2.6
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
filelock==3.0.12
Pillow==8.3.1
platformdirs==2.2.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.7.2
